requirements.txt
# Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==1.4.49
alembic==1.13.1

# API
fastapi==0.95.2
uvicorn==0.27.1
pydantic[email]==1.10.13

# Scheduler (lightweight replacement for Airflow)
APScheduler==3.10.4

# Utils
python-dotenv==1.0.1
httpx==0.27.0
fake-useragent==1.4.0
